requirement.txt
streamlit==1.49.1
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-groq==0.3.7
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.11
chromadb==1.0.20
beautifulsoup4==4.13.5
sentence-transformers==5.1.0
python-dotenv==1.1.1
requests==2.32.5
groq==0.31.0
faiss-cpu==1.12.0

